+++ src/define.js
@@ -173,6 +173,12 @@ export const RAW_VALUE = 'value';
 export const RAW_WHITESPACE = ' ';
 
 /**
+ * Regular expression patterns to validate email addresses.
+ * @type {RegExp}
+ */
+export const RGX_EMAIL = /^[\w-.]+@([\w-]+\.)+[\w-]{2,4}$/g;
+
+/**
  * Regular expression to match non-whitespace.
  * @type {RegExp}
  */